# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Book:
    properties:
      autore:
        type: string
      categoria:
        type: string
      copertina:
        type: string
      genere:
        type: string
      id:
        type: integer
      id_copertina:
        type: integer
      prezzo:
        type: number
      quantita:
        maximum: 5
        minimum: 1
        type: integer
      summary:
        maxLength: 512
        type: string
      titolo:
        maxLength: 255
        type: string
    required:
    - autore
    - categoria
    - copertina
    - genere
    - id_copertina
    - prezzo
    - quantita
    - summary
    - titolo
    type: object
  responses.PaginationDatabase:
    properties:
      page:
        type: integer
      total_pages:
        type: integer
      total_record:
        type: integer
    type: object
  responses.ResponseDatabase:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Book'
        type: array
      paging:
        $ref: '#/definitions/responses.PaginationDatabase'
    type: object
  responses.ResponseErrorJSON:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
host: 192.168.3.6:8000
info:
  contact:
    email: chrisd3v3l0p3r@gmail.com
    name: Chris Developer
  description: This API manage a cart
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Book Store API
  version: "1.0"
paths:
  /book:
    get:
      consumes:
      - application/json
      description: Get details for every book
      parameters:
      - description: Number of the pagination
        in: query
        name: page
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ResponseDatabase'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ResponseErrorJSON'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ResponseErrorJSON'
      summary: Get books
      tags:
      - Book
  /book/:
    post:
      consumes:
      - application/json
      description: Adds new book
      parameters:
      - description: Adds new book
        in: body
        name: models.Book
        required: true
        schema:
          $ref: '#/definitions/models.Book'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Book'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ResponseErrorJSON'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ResponseErrorJSON'
      summary: Adds a book
      tags:
      - Book
  /book/{title}:
    get:
      consumes:
      - application/json
      description: Get details of a book
      parameters:
      - description: Title of the book
        in: path
        name: title
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Book'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ResponseErrorJSON'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ResponseErrorJSON'
      summary: Get a book
      tags:
      - Book
swagger: "2.0"